PythonTMTResearch/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.47.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "schedule>=1.2.2",
    "streamlit>=1.51.0",
    "trafilatura>=2.0.0",
    "twscrape>=0.17.0",
]
